# internal/core/domain/transaction.go
package domain

// Transaction represents a transaction in the blockchain.
type Transaction struct {
	Hash        string `json:"hash"`
	From        string `json:"from"`
	To          string `json:"to"`
	Value       string `json:"value"`
	BlockNumber string `json:"blockNumber"`
}